Renamed references to be consistently named (trace -> tracing)
