+++ packages/react-dom/src/server/ReactPartialRenderer.js
@@ -28,7 +28,7 @@ import {
 } from 'shared/ReactFeatureFlags';
 
 import {
-  REACT_DEBUG_TRACE_MODE_TYPE,
+  REACT_DEBUG_TRACING_MODE_TYPE,
   REACT_FORWARD_REF_TYPE,
   REACT_FRAGMENT_TYPE,
   REACT_STRICT_MODE_TYPE,
@@ -1003,7 +1003,7 @@ class ReactDOMServerRenderer {
       }
 
       switch (elementType) {
-        case REACT_DEBUG_TRACE_MODE_TYPE:
+        case REACT_DEBUG_TRACING_MODE_TYPE:
         case REACT_STRICT_MODE_TYPE:
         case REACT_PROFILER_TYPE:
         case REACT_SUSPENSE_LIST_TYPE:
