+++ packages/react-reconciler/src/ReactTypeOfMode.js
@@ -16,4 +16,4 @@ export const StrictMode = 0b00001;
 export const BlockingMode = 0b00010;
 export const ConcurrentMode = 0b00100;
 export const ProfileMode = 0b01000;
-export const DebugTraceMode = 0b10000;
+export const DebugTracingMode = 0b10000;
