+++ packages/shared/ReactSymbols.js
@@ -29,9 +29,9 @@ export let REACT_FUNDAMENTAL_TYPE = 0xead5;
 export let REACT_RESPONDER_TYPE = 0xead6;
 export let REACT_SCOPE_TYPE = 0xead7;
 export let REACT_OPAQUE_ID_TYPE = 0xeae0;
-export let REACT_DEBUG_TRACE_MODE_TYPE;
+export let REACT_DEBUG_TRACING_MODE_TYPE;
 if (enableDebugTracing) {
-  REACT_DEBUG_TRACE_MODE_TYPE = 0xeae1;
+  REACT_DEBUG_TRACING_MODE_TYPE = 0xeae1;
 }
 
 if (typeof Symbol === 'function' && Symbol.for) {
@@ -55,7 +55,7 @@ if (typeof Symbol === 'function' && Symbol.for) {
   REACT_SCOPE_TYPE = symbolFor('react.scope');
   REACT_OPAQUE_ID_TYPE = symbolFor('react.opaque.id');
   if (enableDebugTracing) {
-    REACT_DEBUG_TRACE_MODE_TYPE = symbolFor('react.debug_trace_mode');
+    REACT_DEBUG_TRACING_MODE_TYPE = symbolFor('react.debug_trace_mode');
   }
 }
